requirements.txt
fastapi==0.83.0
SQLAlchemy==1.4.41
psycopg2-binary==2.9.3
